requirements.txt
discord.py==2.4.0
pandas==2.2.2
art==6.2
googletrans==4.0.0rc1
requests==2.32.3
PyNaCL==1.5.0
yt-dlp==2024.12.6
pytest==8.3.4
pytest-asyncio==0.24.0